+++ libs/api/domains/documents/src/lib/models/v2/actions.ts
@@ -1,0 +1,16 @@
+import { Field, ObjectType } from '@nestjs/graphql'
+
+@ObjectType('DocumentV2Action')
+export class Action {
+  @Field({ nullable: true })
+  title?: string
+
+  @Field({ nullable: true })
+  type?: string
+
+  @Field({ nullable: true })
+  data?: string
+
+  @Field({ nullable: true })
+  icon?: string
+}
