+++ libs/api/domains/documents/src/lib/models/v2/confirmActions.input.ts
@@ -9,4 +9,7 @@ export class DocumentConfirmActionsInput {
 
   @Field(() => Boolean, { nullable: true })
   readonly confirmed?: boolean
+
+  @Field(() => [String], { nullable: true })
+  readonly actions?: string[]
 }
