+++ libs/portals/my-pages/documents/src/components/DocumentLine/DocumentLine.tsx
@@ -25,7 +25,7 @@ import { useDocumentContext } from '../../screens/Overview/DocumentContext'
 import {
   useDocumentConfirmActionsLazyQuery,
   useGetDocumentInboxLineV3LazyQuery,
-} from '../../screens/Overview/Overview.generated'
+} from '../../queries/Overview.generated'
 import { messages } from '../../utils/messages'
 import { FavAndStash } from '../FavAndStash/FavAndStash'
 import UrgentTag from '../UrgentTag/UrgentTag'
@@ -259,8 +259,13 @@ export const DocumentLine: FC<Props> = ({
   }
 
   const confirmActionCaller = (confirmed: boolean | null) => {
+    const actions = documentLine.actions?.map(
+      (action) => `${action.title}: ${action.data}`,
+    )
     confirmAction({
-      variables: { input: { id: documentLine.id, confirmed: confirmed } },
+      variables: {
+        input: { id: documentLine.id, confirmed: confirmed, actions: actions },
+      },
     })
   }
   const unread = !documentLine.opened && !localRead.includes(documentLine.id)
