+++ libs/portals/my-pages/documents/src/hooks/useMailActionV2.ts
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import { toast } from '@island.is/island-ui/core'
 import { useLocale } from '@island.is/localization'
 import { m } from '@island.is/portals/my-pages/core'
-import { useMailActionV2Mutation } from '../screens/Overview/Overview.generated'
+import { useMailActionV2Mutation } from '../queries/Overview.generated'
 import { MailActions } from '../utils/types'
 import { useDocumentList } from './useDocumentList'
 import { messages as docMessages } from '../utils/messages'
