+++ libs/portals/my-pages/documents/src/queries/BatchMailAction.graphql
