+++ libs/portals/my-pages/documents/src/queries/PdfRenderer.graphql
